Bubble Illness

+++ Assets/Scripts/Shrimp/ShrimpStats.cs
@@ -9,7 +9,7 @@ public struct ShrimpStats
     public string name;
     public bool gender;  // True = male, False = female
     public float birthTime;  // The time at the point of the shrimp's birth
-    public int hunger;  // Value from 0-100
+    public float hunger;  // Value from 0-100
     public int illness;  // Value from 0-100
     public int temperament;  // Value from 0-100
